+++ components/core/ui/hexkeyboard/src/main/kotlin/com/flipperdevices/core/ui/hexkeyboard/HexKey.kt
@@ -1,0 +1,8 @@
+package com.flipperdevices.core.ui.hexkeyboard
+
+enum class HexKey(val title: Char) {
+    One('1'), Two('2'), Three('3'), A('A'), B('B'), C('C'),
+    Four('4'), Five('5'), Six('6'), D('D'), E('E'), F('F'),
+    Seven('7'), Eight('8'), Nine('9'), Clear('<'), Ok('K'),
+    Zero('0')
+}